Moved these components to common folder
