+++ src/common/Footer/index.js
@@ -6,7 +6,7 @@ import {
   FooterLinks,
   StyledGitBlack,
   StyledLinkedinIcon,
-} from "./styled";
+} from "./styled"
 
 const Footer = () => {
   return (
@@ -16,7 +16,7 @@ const Footer = () => {
       <FooterDescrtiption>
         I'm always open to new projects whenever I have the time. If you have a
         website, dashboard or mobile app in mind and need some help to make your
-        ideas come to life, feel free to contact me
+        ideas come to life, feel free to contact me.
       </FooterDescrtiption>
       <FooterLinks>
         <StyledGitBlack />
