+++ src/common/Header/index.js
@@ -12,7 +12,7 @@ import {
   HeaderButton,
   HeaderMessageIcon,
 } from "./styled";
-import poster from "../images/poster.png";
+import poster from "../../images/poster.png";
 
 const Header = () => {
   return (
